fixed unused parameter in bot, moved perft to its own file in generator; added files for bitmasks
