+++ include/bitboard/bitmasks.hpp
@@ -1,0 +1,1 @@
+#pragma once
