+++ src/bitboard/bitmasks.cpp
@@ -1,0 +1,3 @@
+#include <pch.hpp>
+
+#include "bitboard/bitmasks.hpp"
